src/requirements.txt
mkdocs==1.6.0
mkdocs-bibtex==2.16.0
mkdocs-drawio==1.6.4
mkdocs-get-deps==0.2.0
mkdocs-glightbox==0.4.0
mkdocs-macros-plugin==1.3.7
mkdocs-material==9.5.29
mkdocs-material-extensions==1.3.1
mkdocs-obsidian-bridge==1.0.4
mkdocs-roamlinks-plugin==0.3.2
mkdocs-section-index==0.3.9
mkdocs-simple-plugin==3.2.0
pybtex==0.24.0
pymdown-extensions==10.8.1
pypandoc==1.13
python-frontmatter==1.1.0
python-markdown-comments==1.1.0
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2024.5.15
tree-sitter==0.23.2
tree-sitter-markdown==0.3.2
tree-sitter-yaml==0.7.0
pytest==8.3.4
oauth2client==4.1.3
gspread==6.1.4
flake8
